seperate out stuff
