+++ dlstreamer-pipeline-server/user_scripts/gvapython/sscape/sscape_adapter.py
@@ -11,7 +11,6 @@ import time
 from collections import defaultdict
 from datetime import datetime
 from uuid import getnode as get_mac
-from pprint import pprint
 
 import cv2
 import ntplib
@@ -20,6 +19,7 @@ import paho.mqtt.client as mqtt
 from pytz import timezone
 
 from utils import publisher_utils as utils
+from car_license_plate_utils import CarLicensePlateProcessor
 
 ROOT_CA = os.environ.get('ROOT_CA', '/run/secrets/certs/scenescape-ca.pem')
 DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%f"
@@ -140,6 +140,9 @@ class PostInferenceDataPublish:
     self.setupMQTT()
     self.metadatagenpolicy = metadatapolicies[metadatagenpolicy]
     self.frame_level_data = {'id': cameraid, 'debug_mac': getMACAddress()}
+    
+    # Initialize car/license plate processor for specific use cases
+    self.car_lp_processor = CarLicensePlateProcessor()
     return
 
   def on_connect(self, client, userdata, flags, rc):
@@ -172,6 +175,8 @@ class PostInferenceDataPublish:
 
   def annotateObjects(self, img):
     objColors = ((0, 0, 255), (255, 128, 128), (207, 83, 294), (31, 156, 238))
+    
+    # Generic object annotation
     for otype, objects in self.frame_level_data['objects'].items():
       if otype == "person":
         cindex = 0
@@ -181,61 +186,18 @@ class PostInferenceDataPublish:
         cindex = 1
       else:
         cindex = 2
-      for obj in objects:
-        topleft_cv = (int(obj['bounding_box_px']['x']), int(obj['bounding_box_px']['y']))
-        bottomright_cv = (int(obj['bounding_box_px']['x'] + obj['bounding_box_px']['width']),
-                        int(obj['bounding_box_px']['y'] + obj['bounding_box_px']['height']))
-        cv2.rectangle(img, topleft_cv, bottomright_cv, objColors[cindex], 4)
-        
-        # Annotate license plates for cars (nested) - draw both bounding box and text
-        if otype == "car" and 'license_plates' in obj:
-          for plate in obj['license_plates']:
-            # Draw license plate bounding box
-            plate_topleft = (int(plate['bounding_box_px']['x']), int(plate['bounding_box_px']['y']))
-            plate_bottomright = (int(plate['bounding_box_px']['x'] + plate['bounding_box_px']['width']),
-                               int(plate['bounding_box_px']['y'] + plate['bounding_box_px']['height']))
-            cv2.rectangle(img, plate_topleft, plate_bottomright, objColors[3], 4)  # Use license plate color
-            
-            # Draw OCR text
-            if 'text' in plate and plate['text']:
-              self.annotatePlate(img, plate['bounding_box_px'], plate['text'])
+      
+      # Handle car objects with special logic via utility
+      if otype == "car":
+        self.car_lp_processor.annotateCarLicensePlates(img, self.frame_level_data['objects'], objColors)
+      else:
+        # Generic annotation for other object types
+        for obj in objects:
+          topleft_cv = (int(obj['bounding_box_px']['x']), int(obj['bounding_box_px']['y']))
+          bottomright_cv = (int(obj['bounding_box_px']['x'] + obj['bounding_box_px']['width']),
+                          int(obj['bounding_box_px']['y'] + obj['bounding_box_px']['height']))
+          cv2.rectangle(img, topleft_cv, bottomright_cv, objColors[cindex], 2)
     return
-
-  def annotatePlate(self, frame, bounds, text):
-    """Annotate license plate text near the license plate bounding box"""
-    # Get an estimate of the size of the text with scale 1
-    scale = 1
-    lsize = cv2.getTextSize(text, cv2.FONT_HERSHEY_SIMPLEX, 1*scale, 5*scale)[0]
-
-    # Then adjust the scale so the text is about twice the length of the plate
-    # (this makes it more or less readable in the annotation without taking too much space)
-    if lsize[0] > 0:
-      scale = scale * 2 * bounds['width'] / lsize[0]
-    lsize = cv2.getTextSize(text, cv2.FONT_HERSHEY_SIMPLEX, 1*scale, int(5*scale))[0]
-
-    start_x = int(bounds['x'] - lsize[0])
-    bottom_y = int(bounds['y'] + 10 + lsize[1])
-    end_x = int(bounds['x'])
-    top_y = int(bounds['y'] + 10)
-    
-    # Check if annotation is within image bounds
-    if self.pointsInsideImage(frame, [[start_x, top_y], [end_x, bottom_y]]):
-      # Draw text with black outline and white fill (for better visibility)
-      cv2.putText(frame, text, (start_x, bottom_y),
-                  cv2.FONT_HERSHEY_SIMPLEX, 1 * scale, (0,0,0), int(5 * scale))
-      cv2.putText(frame, text, (start_x, bottom_y),
-                  cv2.FONT_HERSHEY_SIMPLEX, 1 * scale, (255,255,255), int(2 * scale))
-    return
-
-  def pointsInsideImage(self, frame, img_pts):
-    """Check if points are within image boundaries"""
-    frame_height, frame_width = frame.shape[:2]
-    for point in img_pts:
-      pt_x = int(point[0])
-      pt_y = int(point[1])
-      if pt_x < 0 or pt_x > frame_width or pt_y < 0 or pt_y > frame_height:
-        return False
-    return True
 
   def annotateFPS(self, img, fpsval):
     # code snippet is taken from annotateFPS method in percebro/videoframe.py
@@ -280,114 +242,15 @@ class PostInferenceDataPublish:
         vaobj['id'] = len(objects[otype]) + 1
         objects[otype].append(vaobj)
     
-    # Apply license plate associations after building initial objects
-    self.associateLicensePlates(objects)
+    # Apply domain-specific processing
+    self.applyDomainSpecificProcessing(objects)
     self.frame_level_data['objects'] = objects
 
-  def calculateOverlapScore(self, car_bbox, plate_bbox):
-    """Calculate overlap score between car and license plate bounding boxes"""
-    try:
-      # Extract coordinates
-      car_x1, car_y1 = car_bbox['x'], car_bbox['y']
-      car_x2, car_y2 = car_x1 + car_bbox['width'], car_y1 + car_bbox['height']
-      
-      plate_x1, plate_y1 = plate_bbox['x'], plate_bbox['y']
-      plate_x2, plate_y2 = plate_x1 + plate_bbox['width'], plate_y1 + plate_bbox['height']
-      
-      # Check if plate center is within expanded car bounds
-      plate_center_x = (plate_x1 + plate_x2) / 2
-      plate_center_y = (plate_y1 + plate_y2) / 2
-      
-      # Expand car bounds by 15% to account for license plates on bumpers
-      car_width = car_x2 - car_x1
-      car_height = car_y2 - car_y1
-      margin_x = car_width * 0.15
-      margin_y = car_height * 0.15
-      
-      expanded_car_x1 = car_x1 - margin_x
-      expanded_car_y1 = car_y1 - margin_y
-      expanded_car_x2 = car_x2 + margin_x
-      expanded_car_y2 = car_y2 + margin_y
-      
-      # Check if plate center is within expanded car
-      if (expanded_car_x1 <= plate_center_x <= expanded_car_x2 and
-          expanded_car_y1 <= plate_center_y <= expanded_car_y2):
-        # Calculate distance-based score (closer = higher score)
-        car_center_x = (car_x1 + car_x2) / 2
-        car_center_y = (car_y1 + car_y2) / 2
-        distance = ((plate_center_x - car_center_x) ** 2 + (plate_center_y - car_center_y) ** 2) ** 0.5
-        # Normalize distance by car diagonal
-        car_diagonal = (car_width ** 2 + car_height ** 2) ** 0.5
-        if car_diagonal > 0:
-          normalized_distance = distance / car_diagonal
-          return max(0, 1.0 - normalized_distance)
-      
-      return 0.0
-    except Exception as e:
-      print(f"Error calculating overlap score: {e}")
-      return 0.0
-
-  def associateLicensePlates(self, objects):
-    """Create associations between cars and license plates"""
-    cars = objects.get('car', [])
-    license_plates = objects.get('license_plate', [])
-    
-    if not cars or not license_plates:
-      return
-    
-    print(f"Associating {len(cars)} cars with {len(license_plates)} license plates")
-    
-    used_plates = set()
-    associations_created = 0
-    
-    for car_idx, car in enumerate(cars):
-      car_bbox = car['bounding_box_px']
-      associated_plates = []
-      
-      # Find best matching license plates
-      plate_scores = []
-      for plate_idx, plate in enumerate(license_plates):
-        if plate_idx in used_plates:
-          continue
-          
-        plate_bbox = plate['bounding_box_px']
-        score = self.calculateOverlapScore(car_bbox, plate_bbox)
-        if score > 0.1:  # Minimum threshold
-          plate_scores.append((score, plate_idx, plate))
-      
-      # Sort by score and take the best matches
-      plate_scores.sort(reverse=True, key=lambda x: x[0])
-      
-      for score, plate_idx, plate in plate_scores[:2]:  # Max 2 plates per car
-        # Extract OCR text
-        plate_info = {
-          'bounding_box_px': plate['bounding_box_px'],
-          'confidence': plate['confidence'],
-          'text': plate['text']
-        }
-        associated_plates.append(plate_info)
-        used_plates.add(plate_idx)
-        associations_created += 1
-        
-        print(f"Associated car {car_idx} with license plate {plate_idx} (score: {score:.2f}, text: '{plate['text']}')")
-      
-      # Add license plates to car object
-      if associated_plates:
-        car['license_plates'] = associated_plates
-    
-    # Remove associated license plates from the main objects list
-    if used_plates:
-      remaining_plates = [plate for idx, plate in enumerate(license_plates) if idx not in used_plates]
-      if remaining_plates:
-        objects['license_plate'] = remaining_plates
-        print(f"Kept {len(remaining_plates)} unassociated license plates in main objects list")
-      else:
-        # Remove license_plate category entirely if all plates were associated
-        if 'license_plate' in objects:
-          del objects['license_plate']
-        print("All license plates were associated with cars, removed license_plate category from main objects list")
-    
-    print(f"Created {associations_created} license plate associations")
+  def applyDomainSpecificProcessing(self, objects):
+    """Apply domain-specific processing like car-license plate associations"""
+    # Check if we have cars and license plates - if so, apply association logic
+    if 'car' in objects and 'license_plate' in objects:
+      self.car_lp_processor.associateLicensePlates(objects)
 
   def processFrame(self, frame):
     if self.client.is_connected():
@@ -408,7 +271,7 @@ class PostInferenceDataPublish:
           self.buildImgData(imgdatadict, frame, False)
         self.client.publish(f"scenescape/image/calibration/camera/{self.cameraid}", json.dumps(imgdatadict))
         self.is_publish_calibration_image = False
-
+      self.frame_level_data['number_of_objects'] = sum(len(v) for v in self.frame_level_data['objects'].values())
       self.client.publish(f"scenescape/data/camera/{self.cameraid}", json.dumps(self.frame_level_data))
       frame.add_message(json.dumps(self.frame_level_data))
     return True
